+++ src/LIN3S/AdminBundle/Resources/private/js/lib/image-popup.js
@@ -1,0 +1,36 @@
+/*
+ * This file is part of the Admin Bundle.
+ *
+ * Copyright (c) 2015-2016 LIN3S <info@lin3s.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Beñat Espiña <benatespina@gmail.com>
+ */
+
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _jquery = require('jquery');
+
+var _jquery2 = _interopRequireDefault(_jquery);
+
+require('magnific-popup');
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+var ImagePopup = function ImagePopup(el) {
+  _classCallCheck(this, ImagePopup);
+
+  (0, _jquery2.default)(el).magnificPopup({
+    type: 'image'
+  });
+};
+
+exports.default = ImagePopup;