+++ src/LIN3S/AdminBundle/Resources/private/js/lib/index.js
@@ -15,7 +15,11 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.EventPublisher = exports.tableListAction = exports.panel = exports.menu = exports.formCollectionType = exports.filter = undefined;
+exports.EventPublisher = exports.tableListAction = exports.panel = exports.ImagePopup = exports.menu = exports.formCollectionType = exports.filter = exports.ConfirmDialog = undefined;
+
+var _confirmDialog = require('./confirm-dialog');
+
+var _confirmDialog2 = _interopRequireDefault(_confirmDialog);
 
 var _filter = require('./filter');
 
@@ -24,6 +28,10 @@ var _filter2 = _interopRequireDefault(_filter);
 var _formCollectionType = require('./form-collection-type');
 
 var _formCollectionType2 = _interopRequireDefault(_formCollectionType);
+
+var _imagePopup = require('./image-popup');
+
+var _imagePopup2 = _interopRequireDefault(_imagePopup);
 
 var _menu = require('./menu');
 
@@ -41,9 +49,11 @@ var _lin3sEventBus = require('lin3s-event-bus');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
+exports.ConfirmDialog = _confirmDialog2.default;
 exports.filter = _filter2.default;
 exports.formCollectionType = _formCollectionType2.default;
 exports.menu = _menu2.default;
+exports.ImagePopup = _imagePopup2.default;
 exports.panel = _panel2.default;
 exports.tableListAction = _tableListAction2.default;
 exports.EventPublisher = _lin3sEventBus.EventPublisher;